Scaffold add esmInputs working
